+++ src/Checkers/Artisan.php
@@ -21,7 +21,7 @@ class Artisan extends Base
     /**
      * @return bool
      */
-    protected function executeAndCheck(): bool
+    protected function executeAndCheck()
     {
         $this->executeArtisan();
 
@@ -31,7 +31,7 @@ class Artisan extends Base
     /**
      * @return bool
      */
-    protected function checkArtisanOutput(): bool
+    protected function checkArtisanOutput()
     {
         $output = IlluminateArtisan::output();
 
@@ -39,7 +39,7 @@ class Artisan extends Base
             $output && preg_match("|{$this->target->shouldReturn}|", $output);
     }
 
-    protected function executeArtisan(): void
+    protected function executeArtisan()
     {
         IlluminateArtisan::call(
             $this->target->command['name'],