+++ CHANGELOG.md
@@ -1,5 +1,49 @@
 # Changelog
 
+## 0.9.13 - 2018-09-24
+### Fixed
+- Fix Result status in Security Checker
+
+## 0.9.12 - 2018-09-24
+### Fixed
+- Fix content types for javascript import
+
+## 0.9.11 - 2018-09-23
+### Changed
+- Notifications are disabled by default
+### Fixed
+- Uptime checker
+- Use full namespaces for Laravel façades
+
+## 0.9.10 - 2018-09-22
+### Added
+- Input to filter resources
+### Changed
+- Flush cache when clicking to refresh one resource in the panel
+- Binaries (composer, ping) are now configurable
+- Added config key 'services' to configure services binaries
+- Checker Ping updated, please review configuration for Latency resources
+- Checker Composer updated, please review configuration for PackagesUpToDate resources
+### Fix
+- Laravel 5.5 support
+
+## 0.9.9 - 2018-09-19
+## 0.9.8 - 2018-09-19
+## 0.9.7 - 2018-09-17
+## 0.9.6 - 2018-09-14
+### Fixed
+- Some minor bugs
+
+## 0.9.5 - 2018-09-12
+### Changed
+- Complete refactor
+- Can now create many targets inside resources
+- Resource charts
+- VueJS Panel
+- Allow users to click to refresh one item
+- Store data on database to plot charts
+- Lots of new Checkers and resources
+
 ## 0.4.0 - 2017-12-18
 ### Changed
 - User will need change config files and surround expressions ({{ <expression> }}) with "" (breaking change)